fix: build against libalpm v15.x (pacman 7.0.0)
arch-audit must use alpm.rs v4.0 for building against libalpm v15.x
(provided by pacman 7.0.0). Upstream hasn't bumped this yet, and I have
found no way to force the transitive dependency purely from our side.

There is a closed merge request[^1] that bumps the dependency, but Cargo
can't use it for the lack of any Git ref pointing to it. So for now, we
have to use a local patch. The official arch-audit Arch package does the
same.[^2]

For transparency, the submodule is just upstream's main with the MR's
commit applied on top.

[^1]: https://gitlab.archlinux.org/archlinux/arch-audit/-/merge_requests/20
[^2]: https://gitlab.archlinux.org/archlinux/packaging/packages/arch-audit/-/blob/c77a0ee1/pacman-7.patch